+++ .changeset/config.json
@@ -1,0 +1,4 @@
+{
+  "changelog": ".changeset/get-changelog-entry",
+  "commit": false
+}
