+++ .changeset/getChangelogEntry.js
@@ -1,0 +1,3 @@
+export default () => {
+  throw new Error("we do not have anything to write changelogs yet");
+};
