+++ packages/apply-release-plan/src/test-utils/simple-get-changelog-entry.ts
@@ -4,7 +4,7 @@ import {
   ComprehensiveRelease,
   VersionType,
   NewChangeset
-} from "@changesets/types/src";
+} from "@changesets/types";
 import { RelevantChangesets } from "../types";
 
 async function getReleaseLine(changeset: NewChangeset) {
