+++ packages/cli/package.json
@@ -20,6 +20,9 @@
     "@types/semver": "^6.0.0",
     "@types/uuid": "^3.4.4",
     "@changesets/git": "^0.0.0",
+    "@changesets/types": "^0.0.0",
+    "@changesets/config": "^0.0.0",
+    "@changesets/parse": "^0.0.0",
     "boxen": "^1.3.0",
     "chalk": "^2.1.0",
     "cli-table": "^0.3.1",
