+++ packages/cli/src/commands/publish/__tests__/releaseCommand.test.ts
