+++ packages/cli/src/commands/version/__tests__/consume.js
