+++ packages/cli/src/commands/version/index.js
