+++ types/human-id.d.ts
@@ -1,0 +1,8 @@
+declare module "human-id" {
+  type humanId = () => string;
+
+  export default function(options: {
+    separator: string;
+    capitalize: boolean;
+  }): string;
+}
